requirements.txt
alpaca-py
python-dotenv
yfinance
openai
ta
